requirements.txt
addict==2.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
complexPyTorch==0.4
contourpy==1.1.1
cycler==0.12.1
einops==0.8.1
filelock==3.16.1
fonttools==4.56.0
fsspec==2025.2.0
fvcore==0.1.5.post20221221
h5py==3.11.0
hdf5storage==0.1.19
hf-xet==1.1.2
huggingface-hub==0.32.0
idna==3.10
imageio==2.35.1
importlib_resources==6.4.5
iopath==0.1.10
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mmcv==2.0.0
mmengine==0.10.7
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pip==24.2
platformdirs==4.3.6
portalocker==3.0.0
psutil==7.0.0
pycocotools==2.0.7
Pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytorch-wavelets==1.3.0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
setuptools==75.1.0
six==1.17.0
sympy==1.13.3
tabulate==0.9.0
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.5.4
tokenizers==0.20.3
tomli==2.2.1
torch==2.4.1
torchaudio==2.4.1
torchinfo==1.8.0
torchvision==0.19.1
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
wheel==0.44.0
yacs==0.1.8
yapf==0.43.0
zipp==3.20.2
